Several small improvements
Found missing reading update function call in the maxon specialization, fixed it.
Added the option for changing thread frequency for the worker thread in the yaml file.

TODO:
Add the r/w access, mapping metadata, etc for the OD registers as comments in front of the
register entries. Also mention information about repetitive OD objects detected in the ESI file. This should be mentioned as comments at the end of object #define.
Find a way to add a mapping between integers and mode of operation enums of the respective devices in order to allow changing the mode of operation on the fly.