[Maxon SDK] Major changes
Added new parameters for motor types, and max gear input speeds.
The idea was to handle all the register limits properly since the failure to write into the velocity registers is primarily because of the unit conversions leading to values which exceed the maximum permissible values by the registers. Dependencies needed to be taken care of in ConfigurationParams.cpp.