[Maxon and Nanotec SDK] Features
Added enums for integer values of modes to the command msgs for ease of use.
Added the EthercatDeviceUtils files to allow for more customizations and vendor specific functions outside od the worker thread to keep the code clean.
Added error codes and registers to the Nanotec TxPDOs for error feedback when in fault.
Modified the test trajectory generator python script for a few more parameters and to serve as an example of how to use the mode name enums in the message file.