[Ethercat ROS] Added object factory based registration
Allows for a direct and easy interface for addition of new device types.
The device configurator implementations don't need vendor type checking now since it is done by the object factory.
This allows for easier implmentation of conditional compilation dependent on the availability of device SDKs since there are far fewer lines to add the conditions to. All of them are just in the EthercatDeviceRos.hpp file and there all of the device specific code is stacked together so the conditional compilations become far easier to pull off.