Successful maxon tests, removed pragma once from EthercatDeviceRos.cpp
Maxon was found to run properly when the master cycle time was higher than the interpolation time of maxon. The RxPDO queue overflow error wasn't encountered in this scenario.
pragma once was removed from EthercatDeviceRos.cpp because it may create issues if devices are defined in separate cpp files in the future.
The next task will be to define the device classes in separate files from the EthercatDeviceRos.hpp file for a cleaner interface which will also be easier to describe through the documentation.