[Ethercat SDK] Class registration added
Partial specialization based solution will lead to a lot of code duplication so the proposed improvements for the next version are as follows:
Make the base class templated for the device object since we can safely assume it to be derived from ether EthercatDevice class anyways. And the typecasting is the only reason we need it.
And then shift to inheritence architecture instead of class specializations. This way any inheritence should be from complete template specifications of the classes, and explicit specializations of members should preserve the current idea of changing the worker based on the device and the message specific callbacks should still be possible.